backend/requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.8.0
django>=5.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
psycopg2-binary
dj-database-url
gunicorn
whitenoise
django-cors-headers
requests
uvicorn==0.27.1
gevent>=23.9.1
websockets==12.0